# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_log_api_logs_upload_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_log_api_logs_upload_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StreamIngestBody": {
        "description": "Push logs from a SIEM/forwarder without multipart file upload.",
        "properties": {
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "lines": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lines"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "source": {
            "default": "webhook",
            "title": "Source",
            "type": "string"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "title": "StreamIngestBody",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Agentic Cybersecurity Threat Intelligence & Incident Response Advisor",
    "title": "ACTIRA API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/logs/ingest": {
      "post": {
        "description": "Realtime JSON ingest for SIEM / rsyslog / custom forwarders.\n\nAuth: header `X-Ingest-Key: <INGEST_API_KEY>` or `Authorization: Bearer <jwt>`.\n\nBody examples:\n  {\"text\": \"raw log lines...\\n\", \"source\": \"rsyslog\"}\n  {\"lines\": [\"line1\", \"line2\"], \"filename\": \"firewall.log\"}",
        "operationId": "ingest_stream_json_api_logs_ingest_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Ingest-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ingest-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamIngestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Stream Json",
        "tags": [
          "logs"
        ]
      }
    },
    "/api/logs/ingest/raw": {
      "post": {
        "description": "Realtime raw-body ingest (text/plain or application/octet-stream).\n\nUseful for syslog-ng / fluent-bit HTTP output plugins.",
        "operationId": "ingest_stream_raw_api_logs_ingest_raw_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Ingest-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ingest-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Log-Source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Log-Source"
            }
          },
          {
            "in": "header",
            "name": "X-Log-Filename",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Log-Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Stream Raw",
        "tags": [
          "logs"
        ]
      }
    },
    "/api/logs/jobs": {
      "get": {
        "operationId": "list_jobs_api_logs_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Jobs",
        "tags": [
          "logs"
        ]
      }
    },
    "/api/logs/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_logs_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Job",
        "tags": [
          "logs"
        ]
      }
    },
...
    "/api/logs/jobs/{job_id}/resume": {
      "post": {
        "description": "Re-queue a failed or hung pipeline job when the durable payload still exists.\n\nWorker picks it up on the next poll (same process or after restart). Use when\nupload shows stuck mid-phase or failed with a recoverable error.",
        "operationId": "resume_job_api_logs_jobs__job_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resume Job",
        "tags": [
          "logs"
        ]
      }
    },
    "/api/logs/upload": {
      "post": {
        "description": "Single-file upload \u2014 durable queue (A-S7) with BackgroundTasks fallback kick.",
        "operationId": "upload_log_api_logs_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_log_api_logs_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Log",
        "tags": [
          "logs"
        ]
      }
    },
    "/api/logs/upload-batch": {
      "post": {
        "description": "Multi-file / ZIP incident-package upload (durable queue A-S7).\n\n- Accepts up to 20 files (or a single .zip containing many)\n- Auto-detects each file's format (Apache / Syslog / JSON / CSV / CEF / LEEF / plain)\n- Correlates events across all files into a single incident",
        "operationId": "upload_batch_api_logs_upload_batch_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_batch_api_logs_upload_batch_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Batch",
        "tags": [
          "logs"
        ]
      }
    },
...
    "/api/v1/logs/ingest": {
      "post": {
        "description": "Realtime JSON ingest for SIEM / rsyslog / custom forwarders.\n\nAuth: header `X-Ingest-Key: <INGEST_API_KEY>` or `Authorization: Bearer <jwt>`.\n\nBody examples:\n  {\"text\": \"raw log lines...\\n\", \"source\": \"rsyslog\"}\n  {\"lines\": [\"line1\", \"line2\"], \"filename\": \"firewall.log\"}",
        "operationId": "ingest_stream_json_api_v1_logs_ingest_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Ingest-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ingest-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamIngestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Stream Json",
        "tags": [
          "logs"
        ]
      }
    },
    "/api/v1/logs/ingest/raw": {
      "post": {
        "description": "Realtime raw-body ingest (text/plain or application/octet-stream).\n\nUseful for syslog-ng / fluent-bit HTTP output plugins.",
        "operationId": "ingest_stream_raw_api_v1_logs_ingest_raw_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Ingest-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ingest-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Log-Source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Log-Source"
            }
          },
          {
            "in": "header",
            "name": "X-Log-Filename",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Log-Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Stream Raw",
        "tags": [
          "logs"
        ]
      }
    },
    "/api/v1/logs/jobs": {
      "get": {
        "operationId": "list_jobs_api_v1_logs_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Jobs",
        "tags": [
          "logs"
        ]
      }
    },
    "/api/v1/logs/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_v1_logs_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Job",
        "tags": [
          "logs"
        ]
      }
    },
...
    "/api/v1/logs/jobs/{job_id}/resume": {
      "post": {
        "description": "Re-queue a failed or hung pipeline job when the durable payload still exists.\n\nWorker picks it up on the next poll (same process or after restart). Use when\nupload shows stuck mid-phase or failed with a recoverable error.",
        "operationId": "resume_job_api_v1_logs_jobs__job_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resume Job",
        "tags": [
          "logs"
        ]
      }
    },
    "/api/v1/logs/upload": {
      "post": {
        "description": "Single-file upload \u2014 durable queue (A-S7) with BackgroundTasks fallback kick.",
        "operationId": "upload_log_api_v1_logs_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_log_api_v1_logs_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Log",
        "tags": [
          "logs"
        ]
      }
    },
    "/api/v1/logs/upload-batch": {
      "post": {
        "description": "Multi-file / ZIP incident-package upload (durable queue A-S7).\n\n- Accepts up to 20 files (or a single .zip containing many)\n- Auto-detects each file's format (Apache / Syslog / JSON / CSV / CEF / LEEF / plain)\n- Correlates events across all files into a single incident",
        "operationId": "upload_batch_api_v1_logs_upload_batch_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_batch_api_v1_logs_upload_batch_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Batch",
        "tags": [
          "logs"
        ]
      }
    },
...
  }
}